refactor(client): split group_info_screen into parts
1752 LOC parent down to 305. Seven sibling part files:
- add_member_dialog (240)
- channels_section (165)
- danger_actions (164)
- disappearing_messages (100)
- header_section (369)
- invite_section (61)
- members_section (394)

Public GroupInfoScreen widget + static show unchanged.
flutter analyze clean, all 1921 client tests pass.
