+++ apps/client/lib/src/screens/group_info_screen.dart
@@ -31,6 +31,14 @@ import '../widgets/context_menu/actions/member_actions_registry.dart';
 import '../widgets/context_menu/echo_context_menu.dart';
 import '../widgets/profile_sheets.dart';
 
+part 'group_info_screen/parts/header_section.dart';
+part 'group_info_screen/parts/members_section.dart';
+part 'group_info_screen/parts/channels_section.dart';
+part 'group_info_screen/parts/invite_section.dart';
+part 'group_info_screen/parts/disappearing_messages.dart';
+part 'group_info_screen/parts/danger_actions.dart';
+part 'group_info_screen/parts/add_member_dialog.dart';
+
 const _kJsonHeaders = {'Content-Type': 'application/json'};
 const _kGroupInfoTitle = 'Group Info';
 
@@ -122,642 +130,8 @@ class _GroupInfoScreenState extends ConsumerState<GroupInfoScreen> {
     }
   }
 
-  /// Filters available contacts by fuzzy-search query.
-  List<Contact> _filterAvailableContacts(
-    List<Contact> available,
-    String query,
-  ) {
-    if (query.isEmpty) return available;
-    final scored = <({Contact contact, double score})>[];
-    for (final c in available) {
-      final nameScore = fuzzyScore(query, c.displayName ?? c.username);
-      final handleScore = fuzzyScore(query, c.username);
-      final best = nameScore > handleScore ? nameScore : handleScore;
-      if (best > 0.2) scored.add((contact: c, score: best));
-    }
-    scored.sort((a, b) => b.score.compareTo(a.score));
-    return scored.map((e) => e.contact).toList();
-  }
-
-  /// Builds a dialog for member selection with search functionality.
-  Future<String?> _showAddMemberDialog(List<Contact> available) {
-    final searchController = TextEditingController();
-    return showDialog<String>(
-      context: context,
-      builder: (dialogContext) {
-        return StatefulBuilder(
-          builder: (ctx, setDialogState) {
-            final query = searchController.text.trim();
-            final filtered = _filterAvailableContacts(available, query);
-            final serverUrl = ref.read(serverUrlProvider);
-            return Dialog(
-              backgroundColor: context.surface,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-                side: BorderSide(color: context.border),
-              ),
-              child: ConstrainedBox(
-                constraints: const BoxConstraints(maxWidth: 480),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    // Title and close button
-                    Padding(
-                      padding: const EdgeInsets.all(16),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            'Add Member',
-                            style: TextStyle(
-                              fontSize: 18,
-                              fontWeight: FontWeight.w600,
-                              color: context.textPrimary,
-                            ),
-                          ),
-                          SizedBox(
-                            width: 44,
-                            height: 44,
-                            child: IconButton(
-                              onPressed: () => Navigator.of(ctx).pop(),
-                              icon: Icon(
-                                Icons.close,
-                                size: 20,
-                                color: context.textSecondary,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    // Search field
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 16),
-                      child: TextField(
-                        controller: searchController,
-                        autofocus: true,
-                        decoration: InputDecoration(
-                          hintText: 'Search contacts...',
-                          hintStyle: TextStyle(color: context.textSecondary),
-                          prefixIcon: Icon(
-                            Icons.search,
-                            size: 18,
-                            color: context.textSecondary,
-                          ),
-                          isDense: true,
-                          contentPadding: const EdgeInsets.symmetric(
-                            horizontal: 12,
-                            vertical: 10,
-                          ),
-                          filled: true,
-                          fillColor: context.surface,
-                          enabledBorder: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                            borderSide: BorderSide(color: context.border),
-                          ),
-                          focusedBorder: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                            borderSide: BorderSide(color: context.accent),
-                          ),
-                        ),
-                        style: TextStyle(
-                          color: context.textPrimary,
-                          fontSize: 14,
-                        ),
-                        onChanged: (_) => setDialogState(() {}),
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    Divider(height: 8, color: context.border),
-                    // Contact list
-                    Expanded(
-                      child: filtered.isEmpty
-                          ? Center(
-                              child: Text(
-                                'No contacts found',
-                                style: TextStyle(color: context.textMuted),
-                              ),
-                            )
-                          : ListView.builder(
-                              shrinkWrap: true,
-                              itemCount: filtered.length,
-                              itemBuilder: (listCtx, idx) {
-                                final contact = filtered[idx];
-                                return SizedBox(
-                                  height: 56,
-                                  child: InkWell(
-                                    onTap: () => Navigator.pop(
-                                      dialogContext,
-                                      contact.userId,
-                                    ),
-                                    child: Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        horizontal: 16,
-                                        vertical: 8,
-                                      ),
-                                      child: Row(
-                                        children: [
-                                          buildAvatar(
-                                            name: contact.username,
-                                            radius: 20,
-                                            imageUrl: resolveAvatarUrl(
-                                              contact.avatarUrl,
-                                              serverUrl,
-                                            ),
-                                          ),
-                                          const SizedBox(width: 12),
-                                          Expanded(
-                                            child: Text(
-                                              contact.displayName ??
-                                                  contact.username,
-                                              style: TextStyle(
-                                                fontSize: 14,
-                                                color: context.textPrimary,
-                                              ),
-                                              maxLines: 1,
-                                              overflow: TextOverflow.ellipsis,
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ),
-                                );
-                              },
-                            ),
-                    ),
-                  ],
-                ),
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-
-  /// Sends the add-member request to the server.
-  Future<void> _submitAddMember(String userId) async {
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-
-    final serverUrl = ref.read(serverUrlProvider);
-    try {
-      final response = await http.post(
-        Uri.parse('$serverUrl/api/groups/${widget.conversationId}/members'),
-        headers: {'Authorization': 'Bearer $token', ..._kJsonHeaders},
-        body: jsonEncode({'user_id': userId}),
-      );
-
-      if (response.statusCode == 200 || response.statusCode == 201) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        await _loadGroupInfo(force: true);
-        if (mounted) {
-          ToastService.show(context, 'Member added', type: ToastType.success);
-        }
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _submitAddMember failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to add member',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _addMember() async {
-    // Load contacts for selection
-    await ref.read(contactsProvider.notifier).loadContacts();
-    if (!mounted) return;
-
-    final contacts = ref.read(contactsProvider).contacts;
-    final existingMemberIds =
-        _conversation?.members.map((m) => m.userId).toSet() ?? {};
-    final available = contacts
-        .where((c) => !existingMemberIds.contains(c.userId))
-        .toList();
-
-    if (available.isEmpty) {
-      ToastService.show(
-        context,
-        'All contacts are already in this group',
-        type: ToastType.info,
-      );
-      return;
-    }
-
-    final selected = await _showAddMemberDialog(available);
-    if (selected == null) return;
-
-    await _submitAddMember(selected);
-  }
-
-  Future<void> _deleteGroup() async {
-    final groupName = _conversation?.name ?? 'this group';
-    final confirmed = await showEchoConfirmDialog(
-      context,
-      title: 'Delete Group',
-      content:
-          'Are you sure you want to delete "$groupName"? '
-          'This will permanently remove all messages and members. '
-          'This action cannot be undone.',
-      confirmLabel: 'Delete',
-      destructive: true,
-    );
-    if (!confirmed) return;
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final response = await http.delete(
-        Uri.parse('$serverUrl/api/groups/${widget.conversationId}'),
-        headers: {'Authorization': 'Bearer $token'},
-      );
-      if (response.statusCode == 200 && mounted) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        if (mounted) {
-          Navigator.pop(context);
-          ToastService.show(context, 'Group deleted', type: ToastType.success);
-        }
-      } else if (mounted) {
-        ToastService.show(
-          context,
-          'Only the group owner can delete this group',
-          type: ToastType.error,
-        );
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _deleteGroup failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to delete group',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _leaveGroup() async {
-    final confirmed = await showEchoConfirmDialog(
-      context,
-      title: 'Leave Group',
-      content: 'Are you sure you want to leave this group?',
-      confirmLabel: 'Leave',
-      destructive: true,
-    );
-
-    if (!confirmed) return;
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-
-    final serverUrl = ref.read(serverUrlProvider);
-    try {
-      final response = await http.post(
-        Uri.parse('$serverUrl/api/groups/${widget.conversationId}/leave'),
-        headers: {'Authorization': 'Bearer $token', ..._kJsonHeaders},
-      );
-
-      if ((response.statusCode == 200 || response.statusCode == 204) &&
-          mounted) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        if (mounted) context.go('/home');
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _leaveGroup failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to leave group',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _kickMember(ConversationMember member) async {
-    final confirmed = await showEchoConfirmDialog(
-      context,
-      title: 'Remove Member',
-      content: 'Remove ${member.username} from this group?',
-      confirmLabel: 'Remove',
-      destructive: true,
-    );
-    if (!confirmed) return;
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final response = await http.delete(
-        Uri.parse(
-          '$serverUrl/api/groups/${widget.conversationId}'
-          '/members/${member.userId}',
-        ),
-        headers: {'Authorization': 'Bearer $token'},
-      );
-      if (response.statusCode == 200 && mounted) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        await _loadGroupInfo(force: true);
-        if (mounted) {
-          ToastService.show(
-            context,
-            '${member.username} removed',
-            type: ToastType.success,
-          );
-        }
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _kickMember failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to remove member',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _banMember(ConversationMember member) async {
-    final confirmed = await showEchoConfirmDialog(
-      context,
-      title: 'Ban Member',
-      content:
-          'Ban ${member.username} from this group? '
-          'They will not be able to rejoin.',
-      confirmLabel: 'Ban',
-      destructive: true,
-    );
-    if (!confirmed) return;
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final response = await http.post(
-        Uri.parse(
-          '$serverUrl/api/groups/${widget.conversationId}'
-          '/ban/${member.userId}',
-        ),
-        headers: {'Authorization': 'Bearer $token'},
-      );
-      if (response.statusCode == 200 && mounted) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        await _loadGroupInfo(force: true);
-        if (mounted) {
-          ToastService.show(
-            context,
-            '${member.username} banned',
-            type: ToastType.success,
-          );
-        }
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _banMember failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to ban member',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _saveGroupName() async {
-    final newTitle = _nameController.text.trim();
-    if (newTitle.isEmpty) return;
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final response = await http.put(
-        Uri.parse('$serverUrl/api/groups/${widget.conversationId}'),
-        headers: {'Authorization': 'Bearer $token', ..._kJsonHeaders},
-        body: jsonEncode({'title': newTitle}),
-      );
-      if ((response.statusCode == 200) && mounted) {
-        setState(() => _isEditingName = false);
-        await _loadGroupInfo();
-        await ref.read(conversationsProvider.notifier).loadConversations();
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _saveGroupName failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to update group name',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _saveDescription() async {
-    final newDesc = _descriptionController.text.trim();
-
-    final token = ref.read(authProvider).token;
-    if (token == null) return;
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final response = await http.put(
-        Uri.parse('$serverUrl/api/groups/${widget.conversationId}'),
-        headers: {'Authorization': 'Bearer $token', ..._kJsonHeaders},
-        body: jsonEncode({'description': newDesc}),
-      );
-      if ((response.statusCode == 200) && mounted) {
-        setState(() => _isEditingDescription = false);
-        // Re-fetch conversations from server (which now includes description)
-        // and re-sync our local _conversation field from the refreshed
-        // provider state — without this, the screen kept rendering the
-        // pre-save Conversation object even though the DB was up to date.
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        if (!mounted) return;
-        final refreshed = ref
-            .read(conversationsProvider)
-            .conversations
-            .where((c) => c.id == widget.conversationId)
-            .firstOrNull;
-        if (refreshed != null) {
-          setState(() => _conversation = refreshed);
-        }
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _saveDescription failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to update description',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Future<void> _uploadGroupAvatar() async {
-    final result = await FilePicker.pickFiles(
-      type: FileType.image,
-      withData: true,
-    );
-    if (result == null || result.files.isEmpty) return;
-
-    final file = result.files.first;
-    if (file.bytes == null) return;
-
-    // Show the crop dialog; fall through with original bytes if cancelled.
-    final croppedBytes = mounted
-        ? await showAvatarCropDialog(context, file.bytes!)
-        : null;
-    if (croppedBytes == null) return; // user cancelled
-
-    final serverUrl = ref.read(serverUrlProvider);
-
-    try {
-      final uploader = UploadClient(ref.read(authProvider.notifier));
-      final result = await uploader.uploadFile(
-        serverUrl: serverUrl,
-        path: '/api/groups/${widget.conversationId}/avatar',
-        bytes: croppedBytes,
-        fileName: 'avatar.jpg',
-        mimeType: 'image/jpeg',
-        method: 'PUT',
-        fieldName: 'avatar',
-      );
-      if (!mounted) return;
-
-      if (result.ok) {
-        await ref.read(conversationsProvider.notifier).loadConversations();
-        await _loadGroupInfo(force: true);
-        if (mounted) {
-          ToastService.show(
-            context,
-            'Group avatar updated',
-            type: ToastType.success,
-          );
-        }
-      } else {
-        ToastService.show(
-          context,
-          result.errorMessage ??
-              'Failed to upload avatar (${result.statusCode})',
-          type: ToastType.error,
-        );
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _uploadGroupAvatar failed: $e');
-      if (mounted) {
-        ToastService.show(context, 'Upload error: $e', type: ToastType.error);
-      }
-    }
-  }
-
-  Future<void> _showAddChannelDialog() async {
-    final nameController = TextEditingController();
-    String selectedKind = 'text';
-
-    final result = await showDialog<Map<String, String>>(
-      context: context,
-      builder: (dialogContext) => StatefulBuilder(
-        builder: (ctx, setDialogState) => AlertDialog(
-          title: const Text('Add Channel'),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextField(
-                controller: nameController,
-                autofocus: true,
-                decoration: const InputDecoration(
-                  labelText: 'Channel name',
-                  hintText: 'e.g. general',
-                ),
-              ),
-              const SizedBox(height: 16),
-              DropdownButtonFormField<String>(
-                initialValue: selectedKind,
-                decoration: const InputDecoration(labelText: 'Type'),
-                items: const [
-                  DropdownMenuItem(value: 'text', child: Text('Text')),
-                  DropdownMenuItem(value: 'voice', child: Text('Voice')),
-                ],
-                onChanged: (v) {
-                  if (v != null) setDialogState(() => selectedKind = v);
-                },
-              ),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(dialogContext),
-              child: const Text('Cancel'),
-            ),
-            FilledButton(
-              onPressed: () {
-                final name = nameController.text.trim();
-                if (name.isNotEmpty) {
-                  Navigator.pop(dialogContext, {
-                    'name': name,
-                    'kind': selectedKind,
-                  });
-                }
-              },
-              child: const Text('Create'),
-            ),
-          ],
-        ),
-      ),
-    );
-
-    nameController.dispose();
-    if (result == null) return;
-
-    final success = await ref
-        .read(channelsProvider.notifier)
-        .createChannel(widget.conversationId, result['name']!, result['kind']!);
-    if (mounted) {
-      ToastService.show(
-        context,
-        success ? 'Channel created' : 'Failed to create channel',
-        type: success ? ToastType.success : ToastType.error,
-      );
-    }
-  }
-
-  Future<void> _deleteChannel(String channelId, String channelName) async {
-    final confirmed = await showEchoConfirmDialog(
-      context,
-      title: 'Delete Channel',
-      content: 'Delete channel "$channelName"? This cannot be undone.',
-      confirmLabel: 'Delete',
-      destructive: true,
-    );
-    if (!confirmed) return;
-
-    final success = await ref
-        .read(channelsProvider.notifier)
-        .deleteChannel(widget.conversationId, channelId);
-    if (mounted) {
-      ToastService.show(
-        context,
-        success ? 'Channel deleted' : 'Failed to delete channel',
-        type: success ? ToastType.success : ToastType.error,
-      );
-    }
-  }
-
   // ---------------------------------------------------------------------------
-  // Extracted build helpers
+  // Build
   // ---------------------------------------------------------------------------
 
   Widget _buildLoadingState() {
@@ -773,827 +147,6 @@ class _GroupInfoScreenState extends ConsumerState<GroupInfoScreen> {
       body: const Center(child: Text('Could not load group information')),
     );
   }
-
-  Widget _buildGroupAvatar({required bool isOwnerOrAdmin, String? iconUrl}) {
-    final serverUrl = ref.read(serverUrlProvider);
-    final ticket = ref.read(mediaTicketProvider);
-    final hasIcon = iconUrl != null && iconUrl.isNotEmpty;
-
-    Widget avatar;
-    if (hasIcon) {
-      final fullUrl = ticket != null && ticket.isNotEmpty
-          ? '$serverUrl$iconUrl?ticket=$ticket'
-          : '$serverUrl$iconUrl';
-      avatar = CircleAvatar(
-        radius: 56,
-        backgroundColor: Theme.of(context).colorScheme.tertiary,
-        backgroundImage: NetworkImage(fullUrl),
-        onBackgroundImageError: (_, _) {},
-        child: null,
-      );
-    } else {
-      avatar = CircleAvatar(
-        radius: 56,
-        backgroundColor: Theme.of(context).colorScheme.tertiary,
-        child: const Icon(Icons.group_outlined, size: 40),
-      );
-    }
-
-    return Center(
-      child: GestureDetector(
-        onTap: isOwnerOrAdmin ? _uploadGroupAvatar : null,
-        child: Stack(
-          children: [
-            avatar,
-            if (isOwnerOrAdmin)
-              Positioned(
-                bottom: 0,
-                right: 0,
-                child: Container(
-                  padding: const EdgeInsets.all(4),
-                  decoration: BoxDecoration(
-                    color: Theme.of(context).colorScheme.primary,
-                    shape: BoxShape.circle,
-                  ),
-                  child: Icon(
-                    Icons.camera_alt_outlined,
-                    size: 16,
-                    color: Theme.of(context).colorScheme.onPrimary,
-                  ),
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildGroupNameSection({
-    required String displayName,
-    required bool isOwnerOrAdmin,
-  }) {
-    if (_isEditingName) {
-      return Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 48),
-        child: Row(
-          children: [
-            Expanded(
-              child: TextField(
-                controller: _nameController,
-                autofocus: true,
-                decoration: const InputDecoration(
-                  hintText: 'Group name',
-                  isDense: true,
-                ),
-                onSubmitted: (_) => _saveGroupName(),
-              ),
-            ),
-            IconButton(
-              icon: const Icon(Icons.check),
-              tooltip: 'Save group name',
-              onPressed: _saveGroupName,
-            ),
-            IconButton(
-              icon: const Icon(Icons.close),
-              tooltip: 'Cancel editing',
-              onPressed: () => setState(() => _isEditingName = false),
-            ),
-          ],
-        ),
-      );
-    }
-    return Center(
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(displayName, style: Theme.of(context).textTheme.headlineSmall),
-          if (isOwnerOrAdmin)
-            IconButton(
-              icon: const Icon(Icons.edit_outlined, size: 18),
-              tooltip: 'Edit group name',
-              onPressed: () {
-                _nameController.text = displayName;
-                setState(() => _isEditingName = true);
-              },
-            ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildMemberCount(int count) {
-    final isEncrypted = _conversation?.isEncrypted ?? false;
-    final memberLabel = '$count member${count == 1 ? '' : 's'}';
-    return Center(
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          if (isEncrypted) ...[
-            Icon(Icons.lock, size: 12, color: context.textMuted),
-            const SizedBox(width: 6),
-            Text(
-              'End-to-end encrypted · $memberLabel',
-              style: TextStyle(color: context.textMuted, fontSize: 12),
-            ),
-            const SizedBox(width: 8),
-            // Group E2EE is wired but session-key distribution can stall
-            // ("Securing message…", #434).  Label it so testers don't file
-            // duplicates -- remove once #591 lands server-side enforcement.
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-              decoration: BoxDecoration(
-                color: context.accentLight,
-                borderRadius: BorderRadius.circular(4),
-                border: Border.all(color: context.accent, width: 0.5),
-              ),
-              child: Text(
-                'Experimental',
-                style: TextStyle(
-                  color: context.accent,
-                  fontSize: 10,
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-            ),
-          ] else
-            Text(
-              memberLabel,
-              style: TextStyle(color: context.textMuted, fontSize: 12),
-            ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildDescriptionSection({
-    required Conversation conv,
-    required bool isOwnerOrAdmin,
-  }) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 8, 16, 4),
-          child: Row(
-            children: [
-              Text(
-                'Description',
-                style: Theme.of(context).textTheme.labelLarge?.copyWith(
-                  color: Theme.of(context).colorScheme.primary,
-                ),
-              ),
-              const Spacer(),
-              if (isOwnerOrAdmin && !_isEditingDescription)
-                IconButton(
-                  icon: const Icon(Icons.edit_outlined, size: 18),
-                  tooltip: 'Edit description',
-                  onPressed: () {
-                    _descriptionController.text = conv.description ?? '';
-                    setState(() => _isEditingDescription = true);
-                  },
-                ),
-            ],
-          ),
-        ),
-        if (_isEditingDescription)
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: Row(
-              children: [
-                Expanded(
-                  child: TextField(
-                    controller: _descriptionController,
-                    autofocus: true,
-                    decoration: const InputDecoration(
-                      hintText: 'Group description',
-                      isDense: true,
-                    ),
-                    onSubmitted: (_) => _saveDescription(),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.check),
-                  tooltip: 'Save description',
-                  onPressed: _saveDescription,
-                ),
-                IconButton(
-                  icon: const Icon(Icons.close),
-                  tooltip: 'Cancel editing',
-                  onPressed: () =>
-                      setState(() => _isEditingDescription = false),
-                ),
-              ],
-            ),
-          )
-        else
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: Text(
-              conv.description?.isNotEmpty == true
-                  ? conv.description!
-                  : 'No description',
-              style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                color: conv.description?.isNotEmpty == true
-                    ? Theme.of(context).colorScheme.onSurface
-                    : Theme.of(context).colorScheme.onSurfaceVariant,
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-
-  /// Compact role pill shown beside the member's username.
-  ///
-  /// Color mapping matches the desktop members panel (#769):
-  ///   Owner -> warning amber (rare, high authority)
-  ///   Admin -> accent blue
-
-  /// Trailing "..." affordance on a member row. Routes through
-  /// [EchoContextMenu] so the menu items match right-click and
-  /// long-press paths exactly. Hidden when the menu would be empty
-  /// (target is self with no admin actions on offer).
-  Widget? _buildMemberActions({
-    required ConversationMember member,
-    required bool isOwnerOrAdmin,
-    required bool isMe,
-    required String role,
-  }) {
-    if (isMe) return null;
-    return Builder(
-      builder: (btnContext) => IconButton(
-        icon: Icon(
-          Icons.more_vert,
-          size: 18,
-          color: Theme.of(context).colorScheme.onSurfaceVariant,
-        ),
-        tooltip: 'Member actions',
-        onPressed: () {
-          final box = btnContext.findRenderObject() as RenderBox?;
-          final origin =
-              box?.localToGlobal(Offset(0, box.size.height)) ?? Offset.zero;
-          _openMemberContextMenu(
-            anchor: origin,
-            member: member,
-            role: role,
-            viewerIsAdminOrOwner: isOwnerOrAdmin,
-            isMe: isMe,
-          );
-        },
-      ),
-    );
-  }
-
-  /// Build a [MemberTarget] for [member] and open the centralised
-  /// context menu at [anchor]. Used by all three triggers
-  /// (right-click on row, long-press on row, "..." button).
-  ///
-  /// PR-4 scope is migration only: Add/Remove Contact + Block stay
-  /// off until contacts_provider grows the matching methods. Unblock
-  /// already exists, so it's wired conditionally on the live
-  /// blocked-users list.
-  void _openMemberContextMenu({
-    required Offset anchor,
-    required ConversationMember member,
-    required String role,
-    required bool viewerIsAdminOrOwner,
-    required bool isMe,
-  }) {
-    final isBlocked = _isMemberBlocked(member.userId);
-    final canModerate = viewerIsAdminOrOwner && !isMe && role != 'owner';
-
-    final target = MemberTarget(
-      userId: member.userId,
-      username: member.username,
-      isSelf: isMe,
-      targetIsOwner: role == 'owner',
-      viewerIsAdminOrOwner: viewerIsAdminOrOwner,
-      onViewProfile: () => showUserProfileSheet(context, ref, member.userId),
-      onSendMessage: isMe ? null : () => _openDmWithMember(member),
-      onUnblock: (isMe || !isBlocked)
-          ? null
-          : () => _unblockMember(member.userId),
-      onCopyUsername: () =>
-          _copyToClipboardWithToast(member.username, 'Username copied'),
-      onCopyUserId: () =>
-          _copyToClipboardWithToast(member.userId, 'User ID copied'),
-      onKick: canModerate ? () => _kickMember(member) : null,
-      onBan: canModerate ? () => _banMember(member) : null,
-    );
-
-    EchoContextMenu.open(
-      context: context,
-      target: target,
-      anchor: anchor,
-      model: buildMemberMenu(target),
-    );
-  }
-
-  bool _isMemberBlocked(String userId) {
-    return ref
-        .read(contactsProvider)
-        .blockedUsers
-        .any((u) => u.blockedId == userId);
-  }
-
-  Future<void> _openDmWithMember(ConversationMember member) async {
-    try {
-      final conv = await ref
-          .read(conversationsProvider.notifier)
-          .getOrCreateDm(member.userId, member.username);
-      if (!mounted) return;
-      context.go('/home?conversation=${conv.id}');
-    } catch (_) {
-      if (!mounted) return;
-      ToastService.show(
-        context,
-        'Failed to open conversation',
-        type: ToastType.error,
-      );
-    }
-  }
-
-  Future<void> _unblockMember(String userId) async {
-    await ref.read(contactsProvider.notifier).unblockUser(userId);
-  }
-
-  void _copyToClipboardWithToast(String text, String successMessage) {
-    copyToClipboard(context, text, successMessage: successMessage);
-  }
-
-  /// Single roster row. Matches the desktop members panel styling (#769):
-  /// avatar -> online dot -> role icon + username + role pill, activity line
-  /// underneath ("online" / "away" / "You").
-  Widget _buildMemberTile({
-    required ConversationMember member,
-    required String myUserId,
-    required bool isOwnerOrAdmin,
-  }) {
-    final isMe = member.userId == myUserId;
-    final role = member.role ?? 'member';
-    // "You" trumps the presence label on the row that represents the
-    // viewer themselves; everyone else gets the standard activity line.
-    final String activity;
-    if (isMe) {
-      activity = 'You';
-    } else {
-      final presence = ref.watch(userPresenceProvider(member.userId));
-      activity = presenceLabel(presence.status, isOnline: presence.isOnline);
-    }
-
-    return InkWell(
-      onTap: () {
-        // Tapping a row opens the member's profile sheet, matching the
-        // desktop members panel.
-        // Avoid for self -- you can already see your own settings.
-        if (!isMe) {
-          showUserProfileSheet(context, ref, member.userId);
-        }
-      },
-      onSecondaryTapDown: (details) => _openMemberContextMenu(
-        anchor: details.globalPosition,
-        member: member,
-        role: role,
-        viewerIsAdminOrOwner: isOwnerOrAdmin,
-        isMe: isMe,
-      ),
-      onLongPress: () => _openMemberContextMenu(
-        anchor: Offset.zero,
-        member: member,
-        role: role,
-        viewerIsAdminOrOwner: isOwnerOrAdmin,
-        isMe: isMe,
-      ),
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-        child: Row(
-          children: [
-            // The outer InkWell owns the tap; UserAvatar renders the
-            // avatar + presence dot in one go.
-            UserAvatar(
-              userId: member.userId,
-              username: member.username,
-              avatarUrl: member.avatarUrl,
-              radius: 18,
-              showPresence: true,
-              openProfileOnTap: false,
-            ),
-            const SizedBox(width: 10),
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Row(
-                    children: [
-                      if (role == 'owner' || role == 'admin') ...[
-                        MemberRoleIcon(role: role),
-                        const SizedBox(width: 4),
-                      ],
-                      Flexible(
-                        child: Text(
-                          member.username,
-                          overflow: TextOverflow.ellipsis,
-                          style: TextStyle(
-                            color: context.textPrimary,
-                            fontSize: 14,
-                            fontWeight: FontWeight.w500,
-                          ),
-                        ),
-                      ),
-                      MemberRoleBadge(
-                        role: role,
-                        margin: const EdgeInsets.only(left: 6),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 2),
-                  Text(
-                    activity,
-                    style: TextStyle(color: context.textMuted, fontSize: 11),
-                  ),
-                ],
-              ),
-            ),
-            _buildMemberActions(
-                  member: member,
-                  isOwnerOrAdmin: isOwnerOrAdmin,
-                  isMe: isMe,
-                  role: role,
-                ) ??
-                const SizedBox.shrink(),
-          ],
-        ),
-      ),
-    );
-  }
-
-  /// Member roster grouped by role (Owner / Admin / Members), matching the
-  /// desktop members panel (#769). Each section header is uppercase with a
-  /// muted color + letter-spacing; rows sort alphabetically within a section.
-  List<Widget> _buildMembersSection({
-    required Conversation conv,
-    required String myUserId,
-    required bool isOwnerOrAdmin,
-  }) {
-    int sortByName(ConversationMember a, ConversationMember b) =>
-        a.username.toLowerCase().compareTo(b.username.toLowerCase());
-
-    final owners = conv.members.where((m) => m.role == 'owner').toList()
-      ..sort(sortByName);
-    final admins = conv.members.where((m) => m.role == 'admin').toList()
-      ..sort(sortByName);
-    final regulars =
-        conv.members
-            .where((m) => m.role != 'owner' && m.role != 'admin')
-            .toList()
-          ..sort(sortByName);
-
-    Widget sectionHeader(String label) => Padding(
-      padding: const EdgeInsets.fromLTRB(16, 12, 16, 4),
-      child: Text(
-        label,
-        style: TextStyle(
-          color: context.textMuted,
-          fontSize: 11,
-          fontWeight: FontWeight.w700,
-          letterSpacing: 0.6,
-        ),
-      ),
-    );
-
-    Iterable<Widget> renderGroup(
-      String label,
-      List<ConversationMember> group,
-    ) sync* {
-      if (group.isEmpty) return;
-      yield sectionHeader('$label — ${group.length}');
-      for (final m in group) {
-        yield _buildMemberTile(
-          member: m,
-          myUserId: myUserId,
-          isOwnerOrAdmin: isOwnerOrAdmin,
-        );
-      }
-    }
-
-    return [
-      Padding(
-        padding: const EdgeInsets.fromLTRB(16, 8, 16, 4),
-        child: Row(
-          children: [
-            Text(
-              '${conv.members.length} ${conv.members.length == 1 ? 'member' : 'members'}',
-              style: Theme.of(context).textTheme.labelLarge?.copyWith(
-                color: Theme.of(context).colorScheme.primary,
-              ),
-            ),
-            const Spacer(),
-            IconButton(
-              icon: const Icon(Icons.person_add_outlined),
-              tooltip: 'Add member',
-              onPressed: _addMember,
-            ),
-          ],
-        ),
-      ),
-      ...renderGroup('OWNER', owners),
-      ...renderGroup('ADMIN', admins),
-      ...renderGroup('MEMBERS', regulars),
-    ];
-  }
-
-  List<Widget> _buildChannelsSection() {
-    return [
-      const Divider(),
-      Padding(
-        padding: const EdgeInsets.fromLTRB(16, 8, 16, 4),
-        child: Row(
-          children: [
-            Text(
-              'Channels',
-              style: Theme.of(context).textTheme.labelLarge?.copyWith(
-                color: Theme.of(context).colorScheme.primary,
-              ),
-            ),
-            const Spacer(),
-            IconButton(
-              icon: const Icon(Icons.add_outlined),
-              tooltip: 'Add channel',
-              onPressed: _showAddChannelDialog,
-            ),
-          ],
-        ),
-      ),
-      _buildChannelsList(),
-    ];
-  }
-
-  Widget _buildChannelsList() {
-    final channelsState = ref.watch(channelsProvider);
-    final channels = channelsState.channelsFor(widget.conversationId);
-    if (channels.isEmpty) {
-      return Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16),
-        child: Text(
-          channelsState.isLoadingConversation(widget.conversationId)
-              ? 'Loading channels...'
-              : 'No channels',
-          style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-            color: Theme.of(context).colorScheme.onSurfaceVariant,
-          ),
-        ),
-      );
-    }
-    return Column(
-      children: channels.map((channel) {
-        return ListTile(
-          leading: Icon(
-            channel.isText ? Icons.tag : Icons.headset_mic_outlined,
-            size: 20,
-          ),
-          title: Text(channel.name),
-          subtitle: Text(channel.kind),
-          trailing: IconButton(
-            icon: Icon(
-              Icons.delete_outline,
-              size: 18,
-              color: Theme.of(context).colorScheme.error,
-            ),
-            tooltip: 'Delete channel',
-            onPressed: () => _deleteChannel(channel.id, channel.name),
-          ),
-        );
-      }).toList(),
-    );
-  }
-
-  /// Calls `POST /api/groups/:id/invites` to generate a fresh invite token,
-  /// then copies the returned URL to the clipboard.
-  Future<void> _generateAndCopyInviteLink() async {
-    setState(() => _isGeneratingInvite = true);
-    final serverUrl = ref.read(serverUrlProvider);
-    try {
-      final response = await ref
-          .read(authProvider.notifier)
-          .authenticatedRequest(
-            (token) => http.post(
-              Uri.parse(
-                '$serverUrl/api/groups/${widget.conversationId}/invites',
-              ),
-              headers: {
-                'Authorization': 'Bearer $token',
-                'Content-Type': 'application/json',
-              },
-              body: '{}',
-            ),
-          );
-      if (!mounted) return;
-      if (response.statusCode == 201) {
-        final data = jsonDecode(response.body) as Map<String, dynamic>;
-        final url = data['url'] as String? ?? '';
-        await Clipboard.setData(ClipboardData(text: url));
-        if (mounted) {
-          ToastService.show(
-            context,
-            'Invite link copied to clipboard',
-            type: ToastType.success,
-          );
-        }
-      } else {
-        String msg = 'Failed to generate invite link';
-        try {
-          final body = jsonDecode(response.body) as Map<String, dynamic>;
-          msg = body['error'] as String? ?? msg;
-        } catch (_) {}
-        if (mounted) {
-          ToastService.show(context, msg, type: ToastType.error);
-        }
-      }
-    } catch (_) {
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Could not reach server',
-          type: ToastType.error,
-        );
-      }
-    } finally {
-      if (mounted) setState(() => _isGeneratingInvite = false);
-    }
-  }
-
-  List<Widget> _buildActionButtons({required String myRole}) {
-    return [
-      const Divider(),
-      const SizedBox(height: 16),
-      Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16),
-        child: OutlinedButton.icon(
-          onPressed: _isGeneratingInvite ? null : _generateAndCopyInviteLink,
-          icon: _isGeneratingInvite
-              ? const SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                )
-              : const Icon(Icons.link_outlined),
-          label: const Text('Copy Invite Link'),
-        ),
-      ),
-      const SizedBox(height: 12),
-      Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16),
-        child: OutlinedButton.icon(
-          onPressed: _leaveGroup,
-          icon: const Icon(Icons.logout_outlined),
-          label: const Text('Leave Group'),
-          style: OutlinedButton.styleFrom(
-            foregroundColor: Theme.of(context).colorScheme.error,
-          ),
-        ),
-      ),
-      if (myRole == 'owner') ...[
-        const SizedBox(height: 16),
-        const Divider(),
-        const Padding(
-          padding: EdgeInsets.fromLTRB(16, 8, 16, 4),
-          child: Text(
-            'Danger Zone',
-            style: TextStyle(
-              color: EchoTheme.danger,
-              fontSize: 11,
-              fontWeight: FontWeight.w700,
-              letterSpacing: 0.6,
-            ),
-          ),
-        ),
-        ListTile(
-          onTap: _deleteGroup,
-          leading: const Icon(
-            Icons.delete_forever_outlined,
-            color: EchoTheme.danger,
-          ),
-          title: const Text(
-            'Delete Group',
-            style: TextStyle(
-              color: EchoTheme.danger,
-              fontWeight: FontWeight.w500,
-            ),
-          ),
-          subtitle: Text(
-            'Permanently delete this group and all its messages.',
-            style: TextStyle(
-              color: EchoTheme.danger.withValues(alpha: 0.7),
-              fontSize: 12,
-            ),
-          ),
-        ),
-      ],
-      const SizedBox(height: 24),
-    ];
-  }
-
-  // ---------------------------------------------------------------------------
-  // Disappearing messages
-  // ---------------------------------------------------------------------------
-
-  static const _kTtlOptions = [
-    (label: 'Off', seconds: null),
-    (label: '30 seconds', seconds: 30),
-    (label: '5 minutes', seconds: 300),
-    (label: '1 hour', seconds: 3600),
-    (label: '1 day', seconds: 86400),
-    (label: '1 week', seconds: 604800),
-  ];
-
-  String _ttlLabel(int? seconds) {
-    for (final opt in _kTtlOptions) {
-      if (opt.seconds == seconds) return opt.label;
-    }
-    return '$seconds seconds';
-  }
-
-  Future<void> _setDisappearingTtl(int? seconds) async {
-    final serverUrl = ref.read(serverUrlProvider);
-    try {
-      final response = await ref
-          .read(authProvider.notifier)
-          .authenticatedRequest(
-            (token) => http.put(
-              Uri.parse(
-                '$serverUrl/api/conversations/${widget.conversationId}/disappearing',
-              ),
-              headers: {'Authorization': 'Bearer $token', ..._kJsonHeaders},
-              body: jsonEncode({'ttl_seconds': seconds}),
-            ),
-          );
-      if (!mounted) return;
-      if (response.statusCode == 200) {
-        setState(() => _disappearingTtl = seconds);
-        ToastService.show(
-          context,
-          seconds == null
-              ? 'Disappearing messages turned off'
-              : 'Messages disappear after ${_ttlLabel(seconds)}',
-          type: ToastType.success,
-        );
-      } else {
-        ToastService.show(
-          context,
-          'Failed to update disappearing messages',
-          type: ToastType.error,
-        );
-      }
-    } catch (e) {
-      debugPrint('[GroupInfo] _setDisappearingTtl failed: $e');
-      if (mounted) {
-        ToastService.show(
-          context,
-          'Failed to update disappearing messages',
-          type: ToastType.error,
-        );
-      }
-    }
-  }
-
-  Widget _buildDisappearingSection() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const Divider(),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 8, 16, 4),
-          child: Text(
-            'Disappearing Messages',
-            style: Theme.of(context).textTheme.labelMedium?.copyWith(
-              color: Theme.of(context).colorScheme.primary,
-            ),
-          ),
-        ),
-        ListTile(
-          leading: const Icon(Icons.timer_outlined),
-          title: const Text('Auto-delete after'),
-          trailing: DropdownButton<int?>(
-            value: _disappearingTtl,
-            underline: const SizedBox.shrink(),
-            items: _kTtlOptions.map((opt) {
-              return DropdownMenuItem<int?>(
-                value: opt.seconds,
-                child: Text(opt.label),
-              );
-            }).toList(),
-            onChanged: (v) => _setDisappearingTtl(v),
-          ),
-        ),
-        const SizedBox(height: 4),
-      ],
-    );
-  }
-
-  // ---------------------------------------------------------------------------
-  // Build
-  // ---------------------------------------------------------------------------
 
   @override
   Widget build(BuildContext context) {
